# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Self-Healing Ecosystem Model API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type EcosystemType = (typeof EcosystemType)[keyof typeof EcosystemType];

export const EcosystemType = {
  forest: "forest",
  river: "river",
  grassland: "grassland",
  polar: "polar",
} as const;

export type HealthStatusEnum =
  (typeof HealthStatusEnum)[keyof typeof HealthStatusEnum];

export const HealthStatusEnum = {
  Healthy: "Healthy",
  Stressed: "Stressed",
  Collapsing: "Collapsing",
} as const;

export type SpeciesType = (typeof SpeciesType)[keyof typeof SpeciesType];

export const SpeciesType = {
  plant: "plant",
  herbivore: "herbivore",
  predator: "predator",
} as const;

export interface Species {
  id: string;
  name: string;
  type: SpeciesType;
  initialPopulation: number;
}

export interface CreateEcosystemInput {
  name: string;
  type: EcosystemType;
  /**
   * @minimum -10
   * @maximum 50
   */
  temperature: number;
  /**
   * @minimum 0
   * @maximum 5000
   */
  rainfall: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  pollution: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  deforestationRate: number;
  species: Species[];
}

export type InterventionInputType =
  (typeof InterventionInputType)[keyof typeof InterventionInputType];

export const InterventionInputType = {
  reforestation: "reforestation",
  pollution_reduction: "pollution_reduction",
  species_introduction: "species_introduction",
  habitat_restoration: "habitat_restoration",
  hunting_ban: "hunting_ban",
} as const;

export interface InterventionInput {
  type: InterventionInputType;
  /**
   * @minimum 0
   * @maximum 1
   */
  intensity: number;
}

export type RecommendationPriority =
  (typeof RecommendationPriority)[keyof typeof RecommendationPriority];

export const RecommendationPriority = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  CRITICAL: "CRITICAL",
} as const;

export interface Recommendation {
  action: string;
  priority: RecommendationPriority;
  impact: string;
  timeline: string;
}

export interface MLPredictions {
  healthStatus: HealthStatusEnum;
  /**
   * @minimum 0
   * @maximum 100
   */
  healthScore: number;
  recoveryTimeWeeks: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  collapseRisk: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  stabilityIndex: number;
  /**
   * @minimum 0
   * @maximum 1
   */
  biodiversityIndex: number;
  recommendations: Recommendation[];
}

export type EcosystemInterventionsItem = {
  type: string;
  intensity: number;
  appliedAt: string;
};

export interface Ecosystem {
  id: string;
  name: string;
  type: EcosystemType;
  temperature: number;
  rainfall: number;
  pollution: number;
  deforestationRate: number;
  species: Species[];
  predictions: MLPredictions;
  interventions: EcosystemInterventionsItem[];
  createdAt: string;
  updatedAt: string;
}

export interface SimulationDataPoint {
  week: number;
  plants: number;
  herbivores: number;
  predators: number;
  healthScore: number;
}

export interface SimulationResult {
  id: string;
  ecosystemId: string;
  weeks: number;
  dataPoints: SimulationDataPoint[];
  finalPredictions: MLPredictions;
  runAt: string;
}

export interface SimulationRunInput {
  /**
   * @minimum 1
   * @maximum 520
   */
  weeks?: number;
}